% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_check.R
\name{dl_check}
\alias{dl_check}
\title{A shortcut to download a file and check whether it was previously downloaded}
\usage{
dl_check(url, destfile, ...)
}
\arguments{
\item{url}{URL of the file to be downloaded (see \code{\link[curl:curl_download]{curl::curl_download()}}).}

\item{destfile}{destination file (see \code{\link[curl:curl_download]{curl::curl_download()}})}

\item{...}{further arguments passed to \code{\link[curl:curl_download]{curl::curl_download()}}.}
}
\description{
A shortcut to download a file and check whether it was previously downloaded
}
