% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{extract climate data}
\usage{
extract_data(
  years,
  info,
  geom,
  path = "climateData/",
  outdir = "output/",
  base = "res_",
  quiet = FALSE,
  rm_dir = TRUE
)
}
\arguments{
\item{years}{vector of selected years (numeric).}

\item{info}{a character string specifying the kind of data to be retrieved. One among "bio", "cmi", "mint", "maxt", "pcp", "sg". Default is set to "bio".}

\item{geom}{geometry used for extraction.}

\item{path}{path to the folder where files are stored.}

\item{outdir}{output directory.}

\item{base}{base name of the output files.}

\item{quiet}{Should extra information reported on progress be suppressed?}

\item{rm_dir}{Should data folder be removed?}
}
\description{
extract climate data
}
