% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_noam.R
\name{get_climate_nam_grids}
\alias{get_climate_nam_grids}
\title{Download raster of climate data for North America}
\usage{
get_climate_nam_grids(years, infos = "bio", res = 300, path = ".")
}
\arguments{
\item{years}{vector of selected years (numeric), all years from 1900 until 2015 are available.}

\item{infos}{character string specifying the data to be downloaded. Available data are "bio", "cmi", "mint", "maxt", "pcp" and "sg" (see details).}

\item{res}{resolution of the raster files to be downloaded in arcsec (60 or 300arcsed, i.e. 10km2 or 2km2), default is set to 300.}

\item{path}{path to the folder where files are stored.}
}
\description{
Download raster of climate data for North America
}
\details{
\itemize{
\item \code{bio} : 19 bioclimatic variables,
\item \code{cmi} : monthly Climate Moisture Index,
\item \code{mint}: monthly minimum temperature,
\item \code{maxt}: monthly maximum temperature,
\item \code{pcp} : monthly precipitation,
\item \code{sg}  : 16 additional bioclimatic variables.
}
}
\examples{
\dontrun{
 get_climate_nam_grids(2014:2015)
 get_climate_nam_grids(2000:2001, c("bio", "sg", "pcp"))
}
}
\references{
\url{ftp://ftp.nrcan.gc.ca/pub/outgoing/NAM_grids}
\url{https://climate.weather.gc.ca/}
}
