% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_noam.R
\name{get_climate_noam}
\alias{get_climate_noam}
\title{Retrieve raster of climate data for North America}
\usage{
get_climate_noam(years = 1900:2015, info = c("bio", "cmi", "mint",
  "maxt", "pcp", "sg"), res = 300, path = "climateData/",
  rm_zip = TRUE, quiet = FALSE)
}
\arguments{
\item{years}{vector of selected years (numeric).}

\item{info}{a character string specifying the kind of data to be retrieved. One among "bio", "cmi", "mint", "maxt", "pcp", "sg". Default is set to "bio".}

\item{res}{resolution of the raster files to be downloaded in arcsec (60 or 300arcsed, i.e. 10km2 or 2km2, default is set to 300).}

\item{path}{path to the folder where files are stored.}

\item{rm_zip}{a logical. Should zip files be removed? Default is set to \code{TRUE}.}

\item{quiet}{Should extra information reported on progress be suppressed?}
}
\description{
Retrieve raster of climate data for North America
}
\references{
\url{ftp://ftp.nrcan.gc.ca/pub/outgoing/NAM_grids}
\url{https://climate.weather.gc.ca/}
}
