% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fa_icons.R
\name{get_fa_icons}
\alias{get_fa_icons}
\alias{search_fa_icons}
\title{Retrieve (free) font awesome icons.}
\usage{
get_fa_icons(names, format = c("svg", "png"), res = 128)

search_fa_icons(pattern)
}
\arguments{
\item{names}{icons names}

\item{format}{image format either "svg" (default) or "png".}

\item{res}{resolution for png files (ignored for svg), either 16, 22, 24, 32, 48, 64, 128, or 256.}

\item{pattern}{character string containing a regular expression (see \code{\link[=grep]{grep()}}).}
}
\value{
An objet of class \code{magick-image}, see \code{\link[magick:image_read]{magick::image_read()}}.
}
\description{
Retrieve (free) font awesome icons.
}
\section{Functions}{
\itemize{
\item \code{search_fa_icons}: Search icon names.
}}

\examples{
val <- search_fa_icons("^y")
icons <- get_fa_icons(val, "png", res = 256)
plot(magick::image_append(icons))
}
\references{
\url{https://fontawesome.com/}
\url{https://github.com/encharm/Font-Awesome-SVG-PNG}
}
