% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm.R
\name{get_gadm}
\alias{get_gadm}
\title{Download GADM 3.6 data in all available format.}
\usage{
get_gadm(
  country,
  level = 0,
  path = ".",
  format = c("sf", "sp", "kmz", "shp", "gpkg")
)
}
\arguments{
\item{country}{ISO3 code of a country, use "?" to get the full list of codes.}

\item{level}{administrative level, \code{level = 0} download administrative
boundaries for the entire country. Note that maximum value taken varies
among countries (5 is the maximum value encountered).}

\item{path}{path to the folder where files will be stored.}

\item{format}{either "sf", "sp", "kmz", "shp" or "gpkg" (see details).}
}
\value{
A \code{"SpatialPolygonsDataFrame"} object if \code{format = "sp"}, otherwise an object
of class \code{sf}.
}
\description{
Download GADM 3.6 data in all available format.
}
\details{
\itemize{
\item "sf" downloads a \code{.rds} file containing an object of class \code{SpatialPolygonsDataFrame}.
\item "sp" downloads a \code{.rds} file containing an object of class \code{sf}.
\item "kmz" downloads a \code{.kmz} (Keyhole Markup language Zipped) file.
\item "shp" download an archive zip containing one ESRI shapefiles per level available.
\item "gpkg" downloads an archive zip containing a GeoPackage file including all available levels.
}

Note that if a specific  level does not exist, an HTTP error 404 will be
returned for the first group of files whereas for the second group the
archive will still be downloaded but a error will be thrown when attempting
reading a level that does not exist.
}
\examples{
\dontrun{
can <- get_gadm("CAN", level = 1, format = "sf")
bel <- get_gadm("BEL", format = "gpkg")
}
}
\references{
\url{https://gadm.org/}
}
