% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getextract_climate_noam.R
\name{getextract_climate_noam}
\alias{getextract_climate_noam}
\title{Retrieve raster of climate data for North America and extract values}
\usage{
getextract_climate_noam(
  geom,
  year = 1900,
  info = c("bio", "cmi", "mint", "maxt", "pcp", "sg"),
  res = 300,
  path = "climateData/",
  outdir = "output/",
  rm_zip = TRUE,
  rm_files = TRUE
)
}
\arguments{
\item{geom}{geometry used for extraction.}

\item{year}{selected year (numeric).}

\item{info}{a character string specifying the kind of data to be retrieved. One among "bio", "cmi", "mint", "maxt", "pcp", "sg". Default is set to "bio".}

\item{res}{resolution of the raster files to be downloaded in arcsec (60 or 300arcsed, i.e. 10km2 or 2km2, default is set to 300).}

\item{path}{path to the folder where files are stored.}

\item{outdir}{output directory.}

\item{rm_zip}{a logical. Should zip files be removed?}

\item{rm_files}{a logical. Should raster files be removed after extraction?}
}
\description{
Retrieve raster of climate data for North America and extract values
}
\references{
\url{ftp://ftp.nrcan.gc.ca/pub/outgoing/NAM_grids}
\url{https://climate.weather.gc.ca/}
}
