% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso3}
\alias{iso3}
\title{ISO 3 code}
\format{
A data frame with 256 rows and 2 columns.
}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}
}
\usage{
iso3
}
\description{
A data set containing iso 3 code taken from \code{raster} (\code{?raster::getData ("ISO3")}).
}
\keyword{datasets}
