% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu_code}
\alias{lu_code}
\title{Feature Attributes for land-use data}
\format{
A data frame with 16 rows and 3 variables:\code{label}, \code{code} and \code{definition}.
}
\source{
\url{https://www.agr.gc.ca/atlas/supportdocument_documentdesupport/aafcLand_Use/en/ISO_19131_Land_Use_1990__2000_2010_Data_Product_Specifications.pdf}
}
\usage{
lu_code
}
\description{
A data set containing labels, code and definition of the land-use categories.
}
\keyword{datasets}
