% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_can_land_use.R
\name{get_can_land_use}
\alias{get_can_land_use}
\title{Download raster of land-use (Canada only)}
\usage{
get_can_land_use(years, zones, path = ".")
}
\arguments{
\item{years}{vector of selected years (numeric) or or several 1990, 2000 or 2010.}

\item{zones}{UTM zones. Relevant zones are 7 to 22 (for Canada).}

\item{path}{path to the folder where files will be stored.}
}
\description{
Download raster of land-use (Canada only)
}
\examples{
\dontrun{
 image(raster(get_can_land_use(1990, 7)))
}
}
\references{
\url{https://open.canada.ca/data/en/dataset/18e3ef1a-497c-40c6-8326-aac1a34a0dec}
}
