% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_nam_grids.R
\name{get_climate_nam_grids}
\alias{get_climate_nam_grids}
\alias{extract_climate_data}
\title{Download raster of climate data for North America}
\usage{
get_climate_nam_grids(years, infos = "bio", res = 300, path = ".")

extract_climate_data(path, geom, pattern = "\\\\.asc$|\\\\.tif$", ...)
}
\arguments{
\item{years}{vector of selected years (numeric), all years from 1900 until 2018 are available.}

\item{infos}{character string specifying the data to be downloaded. Available data are "bio", "cmi", "mint", "maxt", "pcp" and "sg" (see details).}

\item{res}{resolution of the raster files to be downloaded in arcsec (60 or 300arcsec, i.e. 10km2 or 2km2), default is set to 300.}

\item{path}{path to the folder where files are or will be stored.}

\item{geom}{geometry used for extraction.}

\item{pattern}{a regular expression. Only file names which match the regular expression will be returned (see \code{\link[=list.files]{list.files()}}).}

\item{...}{additional parameters to be passed to \code{\link[=list.files]{list.files()}}.}
}
\description{
Download raster of climate data for North America
}
\details{
\itemize{
\item \code{bio} : 19 bioclimatic variables,
\item \code{cmi} : monthly Climate Moisture Index,
\item \code{mint}: monthly minimum temperature,
\item \code{maxt}: monthly maximum temperature,
\item \code{pcp} : monthly precipitation,
\item \code{sg}  : 16 additional bioclimatic variables.
}
}
\section{Functions}{
\itemize{
\item \code{extract_climate_data}: extract data from a nam grid.
}}

\examples{
\dontrun{
 get_climate_nam_grids(2018)
 get_climate_nam_grids(2009:2018, c("maxt", "mint", "pcp"))
 # extract_climate_data("2018", get_gadm('CAN', 1)[9, ], full.names = TRUE)
}
\dontrun{
can <- get_gadm("CAN", level = 1, format = "sf")  
get_climate_nam_grids(1990:1991, c("bio")) 
geom <- readRDS("gadm36_CAN_1_sf.rds")
extract_climate_data("1990", geom)
}
}
\references{
\url{ftp://ftp.nrcan.gc.ca/pub/outgoing/NAM_grids}
\url{https://climate.weather.gc.ca/}
}
